+++ tests/discrete_mechanisms/discrete_mechanisms_test.py
@@ -118,13 +118,6 @@ class CalibrationTest(parameterized.TestCase):
   """Tests that calibration works across mechanisms."""
 
   @parameterized.named_parameters(*_MECHANISMS.items())
-  def test_deprecated_zcdp_calibration(self, mechanism):
-    rng = np.random.default_rng(0)
-    data = _make_skewed_dataset(rng)
-    result = mechanism.calibrate(zcdp_rho=_ZCDP_RHO)(rng, data)
-    self.assertIsInstance(result, common.DiscreteMechanismResult)
-
-  @parameterized.named_parameters(*_MECHANISMS.items())
   def test_zero_epsilon_calibration(self, mechanism):
     rng = np.random.default_rng(0)
     data = _make_skewed_dataset(rng)